requirements.txt
Flask==2.0.1
torch==1.9.0
transformers==4.12.0
Pillow==8.4.0
pymongo